Исправлены кавычки. Создано представление для приветственного экрана
