+++ js/data/game-data.test.js
@@ -1,4 +1,4 @@
-import {assert} from 'chai';
+import {assert} from "chai";
 
 describe(`Array`, () => {
   describe(`#indexOf()`, () => {
