+++ js/screens/welcome-screen.js
@@ -1,29 +1,15 @@
 /** @module Приветствие */
 
-import {msToMinutesAndSeconds, pluralize} from '../util';
-import render from '../render';
-import {GAME_SETTINGS} from '../data/game-data';
-import startGame from '../game/start-game';
+import WelcomeView from "../views/welcome-view";
+import startGame from "../game/start-game";
 
-const totalTimeMin = msToMinutesAndSeconds(GAME_SETTINGS.totalTime).minutes;
-const maxMistakes = GAME_SETTINGS.maxMistakes;
+const welcomeScreen = () => {
+  const screen = new WelcomeView();
+  screen.onPlayButtonClick = () => {
+    startGame();
+  };
 
-const welcomeScreenTemplate = `
-<section class="main main--welcome">
-  <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
-  <button class="main-play">Начать игру</button>
-  <h2 class="title main-title">Правила игры</h2>
-  <p class="text main-text">
-    Правила просты&nbsp;— за&nbsp;${totalTimeMin} ${pluralize(totalTimeMin, `minutes`)} ответить на все вопросы.<br>
-    Ошибиться можно ${maxMistakes} ${pluralize(maxMistakes, `attempts`)}.<br>
-    Удачи!
-  </p>
-</section>
-`;
-
-const welcomeScreen = render(welcomeScreenTemplate);
-const playButton = welcomeScreen.querySelector(`.main-play`);
-
-playButton.addEventListener(`click`, () => startGame());
+  return screen.element;
+};
 
 export default welcomeScreen;
