+++ js/views/abstract-view.js
@@ -9,7 +9,7 @@ const render = (markup = ``) => {
   return template.content.firstChild;
 };
 
-class AbstractView {
+export default class AbstractView {
   constructor() {
     if (new.target === AbstractView) {
       throw new Error(`Can't instantiate AbstractView, only concrete one`);
